requirements.txt
streamlit
tensorflow>=2.12
numpy
Pillow
